+++ test/Vault.t.sol
@@ -20,25 +20,27 @@ contract TestLemonJetVault is Test, HelperContract {
 
     function testDepositAndWithdraw() public {
         asset.mint(address(this), 10 ether);
-        uint beforeDepositBalance = asset.balanceOf(address(this));
+        uint256 beforeDepositBalance = asset.balanceOf(address(this));
 
         vault.deposit(1 ether, address(this));
-        vault.redeem(vault.previewWithdraw(vault.previewRedeem(vault.balanceOf(address(this)))), address(this), address(this));
+        vault.redeem(
+            vault.previewWithdraw(vault.previewRedeem(vault.balanceOf(address(this)))), address(this), address(this)
+        );
 
-        uint afterWithdrawBalance = asset.balanceOf(address(this));
+        uint256 afterWithdrawBalance = asset.balanceOf(address(this));
         // assertEq(beforeDepositBalance - afterWithdrawBalance <= Math.mulDiv(...), true);
         console2.log("beforeDepositBalance - afterWithdrawBalance", beforeDepositBalance - afterWithdrawBalance);
     }
 
     function testDepositAndGreaterWithdraw() public {
         asset.mint(address(this), 10 ether);
-        uint beforeDepositBalance = asset.balanceOf(address(this));
+        uint256 beforeDepositBalance = asset.balanceOf(address(this));
         vault.deposit(1 ether, address(this));
 
         asset.mint(address(vault), 10 ether);
 
         vault.withdraw(10 ether, address(this), address(this));
-        uint afterWithdrawBalance = asset.balanceOf(address(this));
+        uint256 afterWithdrawBalance = asset.balanceOf(address(this));
         assertEq(beforeDepositBalance < afterWithdrawBalance, true);
     }
